Restructure project as installable package and refine extraction
